third_party/WebKit/LayoutTests/fast/css/first-letter-hit-test.html

<!doctype html>
<script src="../../resources/js-test.js"></script>
<style>
#tab:first-letter {
  text-transform: uppercase;
}
#tab {
  display: block;
}
</style>
<div>
  <span id="tab"><a id='a'>H</a></span>
</div>

<script>
offset = document.getElementById('tab').getBoundingClientRect();
shouldBe('document.elementFromPoint(offset.left + 5, offset.top + 5).id', "'a'");
</script>
